+++ crest/Assets/Crest/Crest/Scripts/Shapes/FFT/FFTCompute.cs
@@ -349,7 +349,7 @@ namespace Crest
             buf.SetComputeIntParam(_shaderSpectrum, "_Size", _resolution);
             buf.SetComputeFloatParam(_shaderSpectrum, "_WindSpeed", _windSpeed);
             buf.SetComputeFloatParam(_shaderSpectrum, "_Turbulence", _windTurbulence);
-            buf.SetComputeFloatParam(_shaderSpectrum, "_Gravity", _spectrum._gravityScale * Mathf.Abs(Physics.gravity.magnitude));
+            buf.SetComputeFloatParam(_shaderSpectrum, "_Gravity", Mathf.Abs(Physics.gravity.magnitude));
             buf.SetComputeFloatParam(_shaderSpectrum, "_Period", _loopPeriod);
             buf.SetComputeVectorParam(_shaderSpectrum, "_WindDir", new Vector2(Mathf.Cos(_windDirRad), Mathf.Sin(_windDirRad)));
             buf.SetComputeTextureParam(_shaderSpectrum, _kernelSpectrumInit, "_SpectrumControls", _texSpectrumControls);
@@ -362,7 +362,7 @@ namespace Crest
         /// </summary>
         void UpdateSpectrum(CommandBuffer buf, float time)
         {
-            buf.SetComputeFloatParam(_shaderSpectrum, "_Time", time);
+            buf.SetComputeFloatParam(_shaderSpectrum, "_Time", time * _spectrum._gravityScale);
             buf.SetComputeFloatParam(_shaderSpectrum, "_Chop", _spectrum._chop);
             buf.SetComputeFloatParam(_shaderSpectrum, "_Period", _loopPeriod);
             buf.SetComputeTextureParam(_shaderSpectrum, _kernelSpectrumUpdate, "_Init0", _spectrumInit);